requirements.txt
python-chess
torch
tqdm

